Fixes to reduce the number of compile warnings.  Still many more to fix
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@1535 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Utils/MapUtility.cpp
@@ -36,8 +36,8 @@ HVSURFACE	ghvSurface;
 
 extern BOOLEAN gfOverheadMapDirty;
 
-extern int iOffsetHorizontal;
-extern int iOffsetVertical;
+extern UINT16 iOffsetHorizontal;
+extern UINT16 iOffsetVertical;
 
 // Utililty file for sub-sampling/creating our radar screen maps
 // Loops though our maps directory and reads all .map files, subsamples an area, color
